Moved highlight helper into separate module
